requirements.txt
anyio==4.9.0
asgiref==3.9.1
blinker==1.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
python-telegram-bot==22.2
requests==2.32.4
sniffio==1.3.1
tornado==6.5.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
